feat: Updating to latest code in net6 template
