+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.Mobile/MyExtensionsApp.Mobile.csproj
@@ -88,8 +88,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\MyExtensionsApp.Core\MyExtensionsApp.Core.csproj" />
+		<ProjectReference Include="..\MyExtensionsApp\MyExtensionsApp.csproj" />
 	</ItemGroup>
 	
-	<Import Project="..\MyExtensionsApp.Shared\MyExtensionsApp.Shared.projitems" Label="Shared" />
+	<Import Project="..\MyExtensionsApp.UI\MyExtensionsApp.UI.projitems" Label="Shared" />
 </Project>
