+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.Tests/MyExtensionsApp.Tests.csproj
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MyExtensionsApp.Core\MyExtensionsApp.Core.csproj" />
+    <ProjectReference Include="..\MyExtensionsApp\MyExtensionsApp.csproj" />
   </ItemGroup>
 
 </Project>
