+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/App.xaml.cs
@@ -1,0 +1,63 @@
+//-:cnd:noEmit
+
+namespace MyExtensionsApp;
+
+public sealed partial class App : Application
+{
+	private Window? _window;
+
+	public App()
+	{
+		this.InitializeComponent();
+
+#if HAS_UNO || NETFX_CORE
+		this.Suspending += OnSuspending;
+#endif
+	}
+
+	/// <summary>
+	/// Invoked when the application is launched normally by the end user.  Other entry points
+	/// will be used such as when the application is launched to open a specific file.
+	/// </summary>
+	/// <param name="args">Details about the launch request and process.</param>
+	protected async override void OnLaunched(LaunchActivatedEventArgs args)
+	{
+
+#if NET5_0_OR_GREATER && WINDOWS
+		_window = new Window();
+		_window.Activate();
+#else
+		_window = Microsoft.UI.Xaml.Window.Current;
+#endif
+
+		_window.AttachNavigation(Host.Services);
+		_window.Activate();
+
+		await Task.Run(() => Host.StartAsync());
+
+	}
+
+	/// <summary>
+	/// Invoked when Navigation to a certain page fails
+	/// </summary>
+	/// <param name="sender">The Frame which failed navigation</param>
+	/// <param name="e">Details about the navigation failure</param>
+	void OnNavigationFailed(object sender, NavigationFailedEventArgs e)
+	{
+		throw new InvalidOperationException($"Failed to load {e.SourcePageType.FullName}: {e.Exception}");
+	}
+
+	/// <summary>
+	/// Invoked when application execution is being suspended.  Application state is saved
+	/// without knowing whether the application will be terminated or resumed with the contents
+	/// of memory still intact.
+	/// </summary>
+	/// <param name="sender">The source of the suspend request.</param>
+	/// <param name="e">Details about the suspend request.</param>
+	private void OnSuspending(object sender, SuspendingEventArgs e)
+	{
+		var deferral = e.SuspendingOperation.GetDeferral();
+		// TODO: Save application state and stop any background activity
+		deferral.Complete();
+	}
+}
