+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/App.xaml.host.cs
@@ -1,0 +1,78 @@
+//-:cnd:noEmit
+
+using MyExtensionsApp.Configuration;
+
+namespace MyExtensionsApp;
+
+public sealed partial class App : Application
+{
+	private IHost Host { get; } = BuildAppHost();
+
+	private static IHost BuildAppHost()
+	{ 
+		return UnoHost
+				.CreateDefaultBuilder()
+#if DEBUG
+				// Switch to Development environment when running in DEBUG
+				.UseEnvironment(Environments.Development)
+#endif
+
+
+				// Add platform specific log providers
+				.UseLogging(configure: logBuilder =>
+				{
+					// Configure log levels for different categories of logging
+					logBuilder
+							.SetMinimumLevel(LogLevel.Information)
+							.XamlLogLevel(LogLevel.Information)
+							.XamlLayoutLogLevel(LogLevel.Information);
+				})
+
+				.UseConfiguration(configure: configBuilder=>
+					configBuilder
+						.WithEmbeddedAppConfigFile<App>()
+						.RegisterConfiguration<AppConfig>()
+				)
+
+				// Enable localization (see appsettings.json for supported languages)
+				.UseLocalization()
+
+				// Register Json serializers (ISerializer and ISerializer)
+				.UseSerialization()
+
+				// Register services for the application
+				.ConfigureServices(services =>
+				{
+					// TODO: Register your services
+					//services.AddSingleton<IMyService, MyService>();
+				})
+
+
+				// Enable navigation, including registering views and viewmodels
+				.UseNavigation(RegisterRoutes)
+
+				// Add navigation support for toolkit controls such as TabBar and NavigationView
+				.UseToolkitNavigation()
+
+
+				.Build(enableUnoLogging: true);
+
+	}
+	private static void RegisterRoutes(IViewRegistry views, IRouteRegistry routes)
+	{
+		views.Register(
+			new ViewMap<ShellControl,ShellViewModel>(),
+			new ViewMap<MainPage, MainViewModel>(),
+			new ViewMap<SecondPage, SecondViewModel>()
+			);
+
+		routes
+			.Register(
+				new RouteMap("", View: views.FindByViewModel<ShellViewModel>() ,
+						Nested: new RouteMap[]
+						{
+										new RouteMap("Main", View: views.FindByViewModel<MainViewModel>()),
+										new RouteMap("Second", View: views.FindByViewModel<SecondViewModel>()),
+						}));
+	}
+}
