+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/Assets/Icons/close.png
