+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/MyExtensionsAppSettings.cs
@@ -1,0 +1,12 @@
+//-:cnd:noEmit
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MyExtensionsApp
+{
+    public record MyExtensionsAppSettings
+    {
+        public string? LastSearch { get; init; }
+    }
+}
