+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/Strings/en/Resources.resw
