+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/Styles/ColorPaletteOverride.xaml
@@ -1,0 +1,81 @@
+<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+	<ResourceDictionary.ThemeDictionaries>
+		<!-- Light Theme -->
+		<ResourceDictionary x:Key="Light">
+			<!-- Primary -->
+			<Color x:Key="MaterialPrimaryColor">#086AD1</Color>
+			<Color x:Key="MaterialPrimaryVariantDarkColor">#04488C</Color>
+			<Color x:Key="MaterialPrimaryVariantLightColor">#E3EFFB</Color>
+
+			<!-- OnPrimary -->
+			<Color x:Key="MaterialOnPrimaryColor">#FFFFFF</Color>
+
+			<!-- Secondary -->
+			<Color x:Key="MaterialSecondaryColor">#FEB839</Color>
+			<Color x:Key="MaterialSecondaryVariantDarkColor">#FFA502</Color>
+			<Color x:Key="MaterialSecondaryVariantLightColor">#FFD487</Color>
+
+			<!-- OnSecondary -->
+			<Color x:Key="MaterialOnSecondaryColor">#000000</Color>
+
+			<!-- Background -->
+			<Color x:Key="MaterialBackgroundColor">#F0F0F0</Color>
+
+			<!-- OnBackground -->
+			<Color x:Key="MaterialOnBackgroundColor">#000000</Color>
+
+			<!-- Surface -->
+			<Color x:Key="MaterialSurfaceColor">#FFFFFF</Color>
+
+			<!-- OnSurface -->
+			<Color x:Key="MaterialOnSurfaceColor">#000000</Color>
+			
+			<!-- Error -->
+			<Color x:Key="MaterialErrorColor">#B00020</Color>
+
+			<!-- OnError -->
+			<Color x:Key="MaterialOnErrorColor">#FFFFFF</Color>
+		</ResourceDictionary>
+
+		<!-- Dark Theme -->
+		<ResourceDictionary x:Key="Dark">
+			<!-- Primary -->
+			<Color x:Key="MaterialPrimaryColor">#2F81D8</Color>
+			<Color x:Key="MaterialPrimaryVariantDarkColor">#04488C</Color>
+			<Color x:Key="MaterialPrimaryVariantLightColor">#E3EFFB</Color>
+
+			<!-- OnPrimary -->
+			<Color x:Key="MaterialOnPrimaryColor">#FFFFFF</Color>
+
+			<!-- Secondary -->
+			<Color x:Key="MaterialSecondaryColor">#FEB839</Color>
+			<Color x:Key="MaterialSecondaryVariantDarkColor">#FFA502</Color>
+			<Color x:Key="MaterialSecondaryVariantLightColor">#FFD487</Color>
+
+			<!-- OnSecondary -->
+			<Color x:Key="MaterialOnSecondaryColor">#000000</Color>
+
+			<!-- Background -->
+			<Color x:Key="MaterialBackgroundColor">#000000</Color>
+
+			<!-- OnBackground -->
+			<Color x:Key="MaterialOnBackgroundColor">#FFFFFF</Color>
+
+			<!-- Surface -->
+			<Color x:Key="MaterialSurfaceColor">#0F0F0F</Color>
+
+			<!-- OnSurface -->
+			<Color x:Key="MaterialOnSurfaceColor">#FFFFFF</Color>
+			
+			<!-- Error -->
+			<Color x:Key="MaterialErrorColor">#B2213C</Color>
+
+			<!-- OnError -->
+			<Color x:Key="MaterialOnErrorColor">#FFFFFF</Color>
+		</ResourceDictionary>
+
+	</ResourceDictionary.ThemeDictionaries>
+
+</ResourceDictionary>
