+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/appconfig.development.json
@@ -1,0 +1,5 @@
+﻿{
+  "AppInfo": {
+    "Title": "MyExtensionsApp - Dev"
+  }
+}
