+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp.UI/appconfig.json
@@ -1,0 +1,8 @@
+﻿{
+  "AppInfo": {
+    "Title": "MyExtensionsApp"
+  },
+  "LocalizationSettings": {
+    "Cultures": [ "en", "es", "fr" ]
+  }
+}
