+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp/Configuration/AppConfig.cs
@@ -1,0 +1,8 @@
+//-:cnd:noEmit
+
+namespace MyExtensionsApp.Configuration;
+
+public record AppConfig
+{
+	public string? Title { get; init; }
+}
