+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp/GlobalUsings.cs
