+++ src/Uno.Extensions.Templates/content/unoapp-extensions-net6/MyExtensionsApp/_Compat/IsExternalInit.cs
@@ -1,0 +1,11 @@
+//-:cnd:noEmit
+using System;
+using System.Linq;
+
+#if !NET5_0
+namespace System.Runtime.CompilerServices;
+
+internal static class IsExternalInit
+{
+}
+#endif
